Fix : layout_height|width is not definable with android:theme
According to developer.android.com/guide/topics/ui/look-and-feel/themes,
"A style [...] specify the appearance for a single View"
"A theme is [...] applied to an entire app [...], not just an individual view."
So, using "android:theme" to specify layout_width and layout_height does not
make sense and is impossible. That is why, we have to use "style"
instead of "android:theme".

Thanks to this answer : https://stackoverflow.com/a/26281009/8371072