Use a string list for operation
This format is useful to separate the numbers. So, there will be no
problems with the very small or very large numbers like
4.52345423532E-50 where there is a minus. This minus was very confusing
but now, all this number will be put in a case of a list and will be
considered as one number and not a bunch of characters.